Restructure model definitions
